backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0

# Email & Environment
python-dotenv==1.0.0
emails==0.6

# Utilities
python-dateutil==2.8.2

# Machine Learning / Computer Vision
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
numpy==1.26.4
timm==0.9.12
opencv-python==4.8.1.78
ultralytics==8.0.196
facenet-pytorch==2.6.0
pymongo==4.5.0
pyttsx3==2.90
